// app/client/src/components/stories/config/types.tsx
export enum statusType {
  BETA = "beta",
  STABLE = "stable",
  DEPRECATED = "deprecated",
  RELEASE_CANDIDATE = "releaseCandidate",
}

export enum controlType {
  BOOLEAN = "boolean",
  NUMBER = "number",
  RANGE = "range",
  OBJECT = "object",
  ARRAY = "object",
  FILE = "file",
  RADIO = "radio",
  INLINE_RADIO = "inline-radio",
  CHECK = "check",
  INLINE_CHECK = "inline-check",
  SELECT = "select",
  MULTI_SELECT = "multi-select",
  TEXT = "text",
  COLOR = "color",
  DATE = "date",
}